refactor: pull out KeyCmd infrastructure into a separate module
